Refactor the file directory for implement jupyter notebook
