+++ workspace-app/src/containers/App.css
@@ -13,3 +13,7 @@
 .Segment {
   background-color: #222;
 }
+
+a:link {
+  text-decoration: none;
+}
